lambda/facenet/requirements.txt
keras-facenet
pillow
opencv-python-headless
scipy
tensorflow